@@ -30,30 +30,38 @@
     "source": "https://github.com/contao-bootstrap/navbar"
   },
   "require": {
-    "php": "^7.4 || ^8.0",
-    "contao-bootstrap/core": "^2.0",
+    "php": "^8.1",
+    "contao-bootstrap/core": "^3.0",
     "contao-community-alliance/meta-palettes": "^2.0",
-    "contao/core-bundle": "^4.9",
-    "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "netzmacht/contao-toolkit": "^3.7",
-    "netzmacht/html": "^2.0.2",
-    "symfony/config": "^4.4 || ^5.4",
-    "symfony/dependency-injection": "^4.4 || ^5.4",
-    "symfony/http-foundation": "^4.4 || ^5.4",
-    "symfony/http-kernel": "^4.4 || ^5.4",
-    "symfony/routing": "^4.4 || ^5.4",
-    "symfony/translation-contracts": "^1.1 || ^2.0"
+    "contao/core-bundle": "^4.13 || ^5.0",
+    "doctrine/dbal": "^3.4",
+    "mvo/contao-group-widget": "^1.4",
+    "netzmacht/contao-toolkit": "^3.9 || ^4.0",
+    "netzmacht/html": "^2.0.2 || ^3.0",
+    "symfony/config": "^5.4 || ^6.0",
+    "symfony/dependency-injection": "^5.4 || ^6.0",
+    "symfony/http-foundation": "^5.4 || ^6.0",
+    "symfony/http-kernel": "^5.4 || ^6.0",
+    "symfony/routing": "^5.4 || ^6.0",
+    "symfony/translation-contracts": "^1.1 || ^2.0 || ^3.0"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.0",
-    "doctrine/coding-standard": "^9.0.",
+    "doctrine/coding-standard": "^12.0",
     "phpcq/coding-standard": "^2.1.2",
-    "phpcq/plugin-api": "1.x-dev",
     "phpcq/runner-bootstrap": "1.x-dev"
   },
   "conflict": {
     "netzmacht/contao-bootstrap": "*"
   },
+  "repositories": {
+    "contao-bootstrap-core": {
+      "type": "vcs",
+      "url": "https://git.netzmacht.de/contao-bootstrap/core.git"
+    }
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
   "autoload": {
     "psr-4": {
       "ContaoBootstrap\\Navbar\\": "src/"
@@ -61,25 +69,19 @@
   },
   "config": {
     "allow-plugins": {
+      "contao-community-alliance/composer-plugin": true,
       "contao-components/installer": false,
       "contao/manager-plugin": true,
       "dealerdirect/phpcodesniffer-composer-installer": true,
-      "contao-community-alliance/composer-plugin": true
+      "php-http/discovery": true
     },
     "sort-packages": true
   },
   "extra": {
     "branch-alias": {
+      "dev-bootstrap5": "3.0.x-dev",
       "dev-develop": "2.3.x-dev",
       "dev-master": "2.2.x-dev"
-    },
-    "contao": {
-      "transifex": {
-        "languages_cto": "src/Resources/contao/languages",
-        "languages_tx": ".tx",
-        "prefix": "navbar-",
-        "project": "contao-bootstrap"
-      }
     },
     "contao-manager-plugin": "ContaoBootstrap\\Navbar\\ContaoManager\\Plugin"
   }
