@@ -1,17 +1,11 @@
 <?php
 
-/**
- * Contao Bootstrap Navbar.
- *
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Navbar;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class ContaoBootstrapNavbarBundle extends Bundle
+final class ContaoBootstrapNavbarBundle extends Bundle
 {
 }
