@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace ContaoBootstrap\Navbar\Helper;
 
-class DropdownItemHelper extends AbstractItemHelper
+final class DropdownItemHelper extends AbstractItemHelper
 {
     /**
-     * {@inheritdoc}
+     * {@inheritDoc}
      */
     public function __construct(array $item)
     {
